+++ processor/src/test/fixtures/expected/com/example/observable/Arez_WriteOutsideTransactionFromDefaultDefaultObservableModel.java
@@ -14,7 +14,7 @@ import javax.annotation.Nonnull;
 import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
-final class Arez_WriteOutsideTransactionObservablesModel extends WriteOutsideTransactionObservablesModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
+final class Arez_WriteOutsideTransactionFromDefaultDefaultObservableModel extends WriteOutsideTransactionFromDefaultDefaultObservableModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;
@@ -24,12 +24,12 @@ final class Arez_WriteOutsideTransactionObservablesModel extends WriteOutsideTra
 
   private long $$arezd$$_time;
 
-  Arez_WriteOutsideTransactionObservablesModel() {
+  Arez_WriteOutsideTransactionFromDefaultDefaultObservableModel() {
     super();
     final ArezContext $$arezv$$_context = Arez.context();
     final int $$arezv$$_id = ( Arez.areNamesEnabled() || Arez.areRegistriesEnabled() || Arez.areNativeComponentsEnabled() ) ? ++$$arezi$$_nextId : 0;
-    final String $$arezv$$_name = Arez.areNamesEnabled() ? "WriteOutsideTransactionObservablesModel." + $$arezv$$_id : null;
-    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "WriteOutsideTransactionObservablesModel", $$arezv$$_id, $$arezv$$_name, () -> $$arezi$$_nativeComponentPreDispose() ) : null;
+    final String $$arezv$$_name = Arez.areNamesEnabled() ? "WriteOutsideTransactionFromDefaultDefaultObservableModel." + $$arezv$$_id : null;
+    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "WriteOutsideTransactionFromDefaultDefaultObservableModel", $$arezv$$_id, $$arezv$$_name, () -> $$arezi$$_nativeComponentPreDispose() ) : null;
     this.$$arezi$$_kernel = new ComponentKernel( Arez.areZonesEnabled() ? $$arezv$$_context : null, Arez.areNamesEnabled() ? $$arezv$$_name : null, $$arezv$$_id, Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, null, Arez.areNativeComponentsEnabled() ? null : this::$$arezi$$_dispose, null, true, false, false );
     this.$$arez$$_time = $$arezv$$_context.observable( Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".time" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> this.$$arezd$$_time : null, Arez.arePropertyIntrospectorsEnabled() ? v -> this.$$arezd$$_time = v : null );
     this.$$arezi$$_kernel.componentConstructed();
@@ -85,10 +85,6 @@ final class Arez_WriteOutsideTransactionObservablesModel extends WriteOutsideTra
 
   @Override
   public void setTime(final long value) {
-    this.$$arezi$$_kernel.safeSetObservable( Arez.areNamesEnabled() ? this.$$arezi$$_kernel.getName() + ".setTime" : null, () -> this.$$arezi$$_setTime( value ) );
-  }
-
-  private void $$arezi$$_setTime(final long value) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'setTime' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
     }
@@ -112,8 +108,8 @@ final class Arez_WriteOutsideTransactionObservablesModel extends WriteOutsideTra
   @Override
   public final boolean equals(final Object o) {
     if ( Arez.areNativeComponentsEnabled() ) {
-      if ( o instanceof Arez_WriteOutsideTransactionObservablesModel ) {
-        final Arez_WriteOutsideTransactionObservablesModel that = (Arez_WriteOutsideTransactionObservablesModel) o;
+      if ( o instanceof Arez_WriteOutsideTransactionFromDefaultDefaultObservableModel ) {
+        final Arez_WriteOutsideTransactionFromDefaultDefaultObservableModel that = (Arez_WriteOutsideTransactionFromDefaultDefaultObservableModel) o;
         return $$arezi$$_id() == that.$$arezi$$_id();
       } else {
         return false;