+++ processor/src/test/fixtures/input/com/example/observable/WriteOutsideTransactionFromDisabledDefaultObservableModel.java
@@ -1,12 +1,13 @@
 package com.example.observable;
 
 import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
 import arez.annotations.Observable;
 
-@ArezComponent
-abstract class WriteOutsideTransactionObservablesModel
+@ArezComponent( defaultWriteOutsideTransaction = Feature.DISABLE )
+abstract class WriteOutsideTransactionFromDisabledDefaultObservableModel
 {
-  @Observable( writeOutsideTransaction = true )
+  @Observable
   public abstract long getTime();
 
   public abstract void setTime( long value );